added new tables to the database

+++ backend/src/main/java/com/example/backend/repository/RecipeRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.example.backend.entities.Recipe;
 
 @Repository
-public interface RecipeRepository extends CrudRepository<Recipe, Integer> {
+public interface RecipeRepository extends CrudRepository<Recipe, Long> {
     
 }
